+++ .gitignore
@@ -100,5 +100,5 @@ ENV/
 # mypy
 .mypy_cache/
 
-# pychram
+# PyCharm
 .idea
