+++ requirements.txt
@@ -1,0 +1,3 @@
+aiohttp==2.3.6
+cerberus==1.1
+tinydb==3.7.0
